[#67]Sitemap bug. - using only admin panel language in links.
1. If there are multiple languages in one store, the language code
will be added after the products, categories and articles url.
In this way, the search engine will index them separately. It 
will not result in duplication because the content linked to 
the url is different.

For example:

	<url>
		<loc>http://www.g1185.me/1-laptop/6-13-3-macbook-air-apple-z0fs0ll-a.html?language=en_US</loc>
		<lastmod>2013-08-13</lastmod>
		<changefreq>daily</changefreq>
		<priority>0.5</priority>
	</url>
	
		<url>
		<loc>http://www.g1185.me/1-laptop/6-13-3-macbook-air-apple-z0fs0ll-a.html?language=en_US</loc>
		<lastmod>2013-08-13</lastmod>
		<changefreq>daily</changefreq>
		<priority>0.5</priority>
	</url>
	
2. Improve some details for the sitemaps
-- For multiple languages, the tomatocart directory should be 
writable to create the sitemap for each language except english.
So, a notification about this is added in the panel so as to 
let the users notice it.

-- If the tomatocart directory is not writable, a error message
will be alerted.