[#125]<sup> tag isn't filtered in the USPS shipping module.
The <sup> just be a html tag in the response from usps. 
I replaced that tag to empty in the code. It is caused 
by the changes in the response from usps. It means the usps 
change the entity. So, it is a bug now. 

I have fixed it for the next v1.1.8.5.